Disallow deletion of active port bindings
Currently it is possible to delete a active binding via the port
binding API extension. Deleting a active port binding leaves the port
in a state of not being editable anymore.
Each attempt editing the port results in a PortNotFound error.

During a nova live migration a port entered the state described above.
For some reasons nova was not able to set the binding of the new target
host to active, thus the live migration rollback process started.
During the rollback, the newly created binding got deleted (apparently
already set to active) leaving the port with no active binding.
Thus setting old binding information was not possible anymore.

Fixing such a zombie requires to set the status of a binding to active
(either via db directly or port binding extension)

+++ neutron/plugins/ml2/plugin.py
@@ -2883,6 +2883,13 @@ class Ml2Plugin(db_base_plugin_v2.NeutronDbPluginV2,
     @utils.transaction_guard
     @db_api.retry_if_session_inactive()
     def delete_port_binding(self, context, host, port_id):
+        port_db = self._get_port(context, port_id)
+        binding = self._get_binding_for_host(port_db.port_bindings, host)
+        if not binding:
+            raise exc.PortBindingNotFound(port_id=port_id, host=host)
+        if binding.status == const.ACTIVE:
+            raise exc.PortBindingInStatusActive(port_id=port_id, host=host)
+
         ports_obj.PortBinding.delete_objects(context,
                                              host=host,
                                              port_id=port_id)
